refactor(precompiles): move B20Abi out of common/abi/mod.rs
mod.rs now holds only the module doc, module declarations, and their paired
re-exports; the B20Abi enum, its impl, and its tests move to
common/abi/b20_abi.rs. No behavior change -- the type is re-exported from the
same path, so every downstream `use` is unaffected.

Addresses https://github.com/base/base/pull/4221#discussion_r3690770110
